Rename cover/ to latloncover/ to match repo/package name
