+++ setup.cfg
@@ -20,4 +20,4 @@ where=src
 
 [options.entry_points]
 console_scripts =
-    cover = cover.main:main
+    cover = latloncover.main:main
