--- src/cover/__init__.py
@@ -1,2 +1,0 @@
-from cover.classify import add_classifications
-__all__ = ['add_classifications']
