+++ src/latloncover/__init__.py
@@ -1,0 +1,2 @@
+from latloncover.classify import add_classifications
+__all__ = ['add_classifications']
